// lib/enums.ts
export const enum MapiObjectType {
    // Address book container object
    MAPI_ABCONT = 4,
    // Address book object
    MAPI_ADDRBOOK = 2,
    // Message attachment object
    MAPI_ATTACH = 7,
    // Distribution list object
    MAPI_DISTLIST = 8,
    // Folder object
    MAPI_FOLDER = 3,
    // Form object
    MAPI_FORMINFO = 12,
    // Messaging user object
    MAPI_MAILUSER = 6,
    // Message object
    MAPI_MESSAGE = 5,
    // Profile section object
    MAPI_PROFSECT = 9,
    // Session object
    MAPI_SESSION = 11,
    // Status object
    MAPI_STATUS = 10,
    // Message store object
    MAPI_STORE = 1,
}
// ...
export const enum PropertyFlag {
    // If this flag is set for a property, that property MUST NOT be deleted from the .msg file
    // (irrespective of which storage it is contained in) and implementations MUST return an error
    // if any attempt is made to do so. This flag is set in circumstances where the implementation
    // depends on that property always being present in the .msg file once it is written there.
    PROPATTR_MANDATORY = 0x00000001,
    // If this flag is not set on a property, that property MUST NOT be read from the .msg file
    // and implementations MUST return an error if any attempt is made to read it. This flag is
    // set on all properties unless there is an implementation-specific reason to prevent a property
    // from being read from the .msg file.
    PROPATTR_READABLE = 0x00000002,
    // If this flag is not set on a property, that property MUST NOT be modified or deleted and
    // implementations MUST return an error if any attempt is made to do so. This flag is set in
    // circumstances where the implementation depends on the properties being writable.
    PROPATTR_WRITABLE = 0x00000004,
}
// ...
export enum AttachmentType {
    // There is no attachment
    NO_ATTACHMENT = 0x00000000,
    // The  PropertyTags.PR_ATTACH_DATA_BIN property contains the attachment data
    ATTACH_BY_VALUE = 0x00000001,
    // The "PropertyTags.PR_ATTACH_PATHNAME_W" or "PropertyTags.PR_ATTACH_LONG_PATHNAME_W"
    // property contains a fully qualified path identifying the attachment to recipients with access to a common file server
    ATTACH_BY_REFERENCE = 0x0002,
    // The "PropertyTags.PR_ATTACH_PATHNAME_W" or "PropertyTags.PR_ATTACH_LONG_PATHNAME_W"
    // property contains a fully qualified path identifying the attachment
    ATTACH_BY_REF_RESOLVE = 0x0003,
    // The "PropertyTags.PR_ATTACH_PATHNAME_W" or "PropertyTags.PR_ATTACH_LONG_PATHNAME_W"
    // property contains a fully qualified path identifying the attachment
    ATTACH_BY_REF_ONLY = 0x0004,
    // The "PropertyTags.PR_ATTACH_DATA_OBJ" (PidTagAttachDataObject) property contains an embedded object
    // that supports the IMessage interface
    ATTACH_EMBEDDED_MSG = 0x0005,
    // The attachment is an embedded OLE object
    ATTACH_OLE = 0x0006,
}

export const enum AttachmentFlags {
    // Indicates that this attachment is not available to HTML rendering applications and should be ignored in
    // Multipurpose Internet Mail Extensions (MIME) processing.
    ATT_INVISIBLE_IN_HTML = 0x00000001,
    // Indicates that this attachment is not available to applications rendering in Rich Text Format (RTF) and should be
    // ignored by MAPI.
    ATT_INVISIBLE_IN_RTF = 0x00000002,
    // The Attachment object is referenced and rendered within the HTML body of the associated Message object.
    ATT_MHTML_REF = 0x00000004,
}

export const enum StoreSupportMask {
    // The message store supports properties containing ANSI (8-bit) characters.
    STORE_ANSI_OK = 0x00020000,
    // The message store supports attachments (OLE or non-OLE) to messages.
    STORE_ATTACH_OK = 0x00000020,
    // The message store supports categorized views of tables.
    STORE_CATEGORIZE_OK = 0x00000400,
    // The message store supports creation of new messages.
    STORE_CREATE_OK = 0x00000010,
    // Entry identifiers for the objects in the message store are unique, that is, never reused during the life of the
    // store.
    STORE_ENTRYID_UNIQUE = 0x00000001,
    // The message store supports HTML messages, stored in the <see cref="PropertyTags.PR_HTML" /> (PidTagBodyHtml)
    // property. Note that STORE_HTML_OK is not defined in versions of MAPIDEFS.H that are included with Microsoft Exchange
    // 2000 Server and earlier. If your development environment uses a MAPIDEFS.H file that does not include STORE_HTML_OK,
    // use the value 0x00010000 instead.
    STORE_HTML_OK = 0x00010000,
    // In a wrapped PST store, indicates that when a new message arrives at the store, the store does rules and spam
    // filter processing on the message separately. The store calls IMAPISupport==Notify, setting fnevNewMail in the
    // NOTIFICATION structure that is passed as a parameter, and then passes the details of the new message to the
    // listening client. Subsequently, when the listening client receives the notification, it does not process rules on
    // the message.
    STORE_ITEMPROC = 0x00200000,
    // This flag is reserved and should not be used.
    STORE_LOCALSTORE = 0x00080000,
    // The message store supports modification of its existing messages.
    STORE_MODIFY_OK = 0x00000008,
    // The message store supports multivalued properties, guarantees the stability of value order in a multivalued
    // property throughout a save operation, and supports instantiation of multivalued properties in tables.
    STORE_MV_PROPS_OK = 0x00000200,
    // The message store supports notifications.
    STORE_NOTIFY_OK = 0x00000100,
    // The message store supports OLE attachments. The OLE data is accessible through an IStorage interface, such as that
    // available through the PR_ATTACH_DATA_OBJ (PidTagAttachDataObject) property
    STORE_OLE_OK = 0x00000040,
    // The folders in this store are public (multi-user), not private (possibly multi-instance but not multi-user).
    STORE_PUBLIC_FOLDERS = 0x00004000,
    // The MAPI Protocol Handler will not crawl the store, and the store is responsible to push any changes through
    // notifications to the indexer to have messages indexed.
    STORE_PUSHER_OK = 0x00800000,
    // All interfaces for the message store have a read-only access level.
    STORE_READONLY = 0x00000002,
    // The message store supports restrictions.
    STORE_RESTRICTION_OK = 0x00001000,
    // The message store supports Rich Text Format (RTF) messages, usually compressed, and the store itself keeps
    // <see cref="PropertyTags.PR_BODY_W" /> and <see cref="PropertyTags.PR_RTF_COMPRESSED" /> synchronized.
    STORE_RTF_OK = 0x00000800,
    // The message store supports search-results folders.
    STORE_SEARCH_OK = 0x00000004,
    // The message store supports sorting views of tables.
    STORE_SORT_OK = 0x00002000,
    // The message store supports marking a message for submission.
    STORE_SUBMIT_OK = 0x00000080,
    // The message store supports storage of RTF messages in uncompressed form. An uncompressed RTF stream is identified
    // by the value dwMagicUncompressedRTF in the stream header. The dwMagicUncompressedRTF value is defined in the
    // RTFLIB.H file
    STORE_UNCOMPRESSED_RTF = 0x00008000,
    // The message store supports properties containing Unicode characters.
    STORE_UNICODE_OK = 0x00040000,
}

export const StoreSupportMaskConst =
    StoreSupportMask.STORE_ATTACH_OK | // | StoreSupportMask.STORE_CATEGORIZE_OK
    StoreSupportMask.STORE_CREATE_OK | //StoreSupportMask.STORE_ENTRYID_UNIQUE
    StoreSupportMask.STORE_MODIFY_OK | // | StoreSupportMask.STORE_MV_PROPS_OK
    // | StoreSupportMask.STORE_OLE_OK
    // | StoreSupportMask.STORE_RTF_OK
    StoreSupportMask.STORE_HTML_OK |
    StoreSupportMask.STORE_UNICODE_OK